Refactoring the registration enrichment call, by introducing a specific asc route base, and introducing .../asc/registration/enrich endpoint to reuse the multiple step search against strapi.
