+++ ascRoutes/registration/index.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+
+const enrichmentRoute = require('./enrichment');
+
+router.use('/enrich', enrichmentRoute);
+
+module.exports = router;
